+++ NEWS.md
@@ -9,6 +9,7 @@
 4. Fix #34 - Add function `tidy_zero_truncated_poisson()`
 5. Fix #35 - Add function `tidy_zero_truncated_geometric()`
 6. Fix #36 - Add function `tidy_zero_truncated_binomial()`
+7. Fix #37 - Add function `tidy_zero_truncated_negative_binomial()`
 
 ## Fixes and Minor Improvements
 1. Fix #30 - Move `crayon`, `rstudioapi`, and `cli` from Suggests to Imports due to `pillar`